Required GLFloat

+++ source/texture/texture.d
@@ -44,8 +44,8 @@ GLuint addTexture(string fileLocation, bool debugEnabled = false) {
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_BORDER);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_BORDER);
 
-    // Border color is nothing
-    double[4] borderColor = [0,0,0,0];
+    // Border color is nothing - This is a GL REQUIRED float
+    float[4] borderColor = [0,0,0,0];
     glTexParameterfv(GL_TEXTURE_2D, GL_TEXTURE_BORDER_COLOR, borderColor.ptr);
 
     // Add in nearest neighbor texture filtering
